+++ Assets/CommonCore/Core/Scripting/ScriptingTypes.cs
@@ -103,11 +103,6 @@ namespace CommonCore.Scripting
         public override string Message => "A matching script could not be found.";
     }
 
-    public class ScriptAlreadyRegisteredException : Exception
-    {
-        public override string Message => "The script has already been registered.";
-    }
-
     public class ArgumentCoercionException : Exception
     {
         public ArgumentCoercionException(Type sourceType, Type targetType, Exception innerException) : base($"Failed to coerce {sourceType.Name} to {targetType.Name}", innerException)